腾讯云cos桶接入完成

+++ RuoYi-Cloud-master/ruoyi-modules/ruoyi-meeting/src/main/java/com/ruoyi/meeting/controller/MeetingController.java
@@ -4,18 +4,13 @@ import java.util.List;
 import java.io.IOException;
 import javax.servlet.http.HttpServletResponse;
 
+import com.ruoyi.common.core.utils.DateUtils;
 import com.ruoyi.cos.api.RemoteCosService;
+import com.ruoyi.meeting.constant.CosConstant;
 import com.ruoyi.meeting.qo.MeetingInsertQuery;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 import com.ruoyi.common.log.annotation.Log;
 import com.ruoyi.common.log.enums.BusinessType;
 import com.ruoyi.common.security.annotation.RequiresPermissions;
@@ -29,14 +24,13 @@ import org.springframework.web.multipart.MultipartFile;
 
 /**
  * 会议Controller
- * 
+ *
  * @author xiye
  * @date 2024-12-25
  */
 @RestController
 @RequestMapping("/meeting")
-public class MeetingController extends BaseController
-{
+public class MeetingController extends BaseController {
     @Autowired
     private IMeetingService meetingService;
     @Autowired
@@ -47,8 +41,7 @@ public class MeetingController extends BaseController
      */
     @RequiresPermissions("meeting:meeting:list")
     @GetMapping("/list")
-    public TableDataInfo list(Meeting meeting)
-    {
+    public TableDataInfo list(Meeting meeting) {
         startPage();
         List<Meeting> list = meetingService.selectMeetingList(meeting);
         return getDataTable(list);
@@ -60,8 +53,7 @@ public class MeetingController extends BaseController
     @RequiresPermissions("meeting:meeting:export")
     @Log(title = "会议", businessType = BusinessType.EXPORT)
     @PostMapping("/export")
-    public void export(HttpServletResponse response, Meeting meeting)
-    {
+    public void export(HttpServletResponse response, Meeting meeting) {
         List<Meeting> list = meetingService.selectMeetingList(meeting);
         ExcelUtil<Meeting> util = new ExcelUtil<Meeting>(Meeting.class);
         util.exportExcel(response, list, "会议数据");
@@ -72,8 +64,7 @@ public class MeetingController extends BaseController
      */
     @RequiresPermissions("meeting:meeting:query")
     @GetMapping(value = "/{id}")
-    public AjaxResult getInfo(@PathVariable("id") Long id)
-    {
+    public AjaxResult getInfo(@PathVariable("id") Long id) {
         return success(meetingService.selectMeetingById(id));
     }
 
@@ -82,40 +73,60 @@ public class MeetingController extends BaseController
      */
     @RequiresPermissions("meeting:meeting:add")
     @Log(title = "会议", businessType = BusinessType.INSERT)
-    @PostMapping
-    public AjaxResult add(@RequestBody Meeting meeting)
-    {
+    @PostMapping("/add")
+    public AjaxResult add(@RequestBody Meeting meeting) {
+        // 如果传了图片就调用然后更新
+        meeting.setCreateTime(DateUtils.getNowDate());
         return toAjax(meetingService.insertMeeting(meeting));
+    }
+
+    @RequiresPermissions("meeting:meeting:add")
+    @PostMapping("/addImage")
+    public AjaxResult add(@RequestPart(value = "file", required = false) MultipartFile file, @RequestParam("imageId") String imageId) {
+        String url = null;
+        if (file != null) {
+            if (!file.isEmpty()) {
+                AjaxResult ajaxResult = remoteCosService.uploadFileCommon(file, imageId);
+                if (ajaxResult.get("code").toString().equals("200")) {
+                    String filename = file.getOriginalFilename();
+                    String extend = filename.substring(filename.lastIndexOf(".") + 1);
+                    url = CosConstant.COS_PATH + "common/" + imageId + "." + extend;
+                }
+            }
+        }
+        return AjaxResult.success(url);
     }
 
     /**
      * 修改会议
      */
-    // @RequiresPermissions("meeting:meeting:edit")
+    @RequiresPermissions("meeting:meeting:edit")
     @Log(title = "会议", businessType = BusinessType.UPDATE)
     @PostMapping("/edit")
-    public AjaxResult edit(MeetingInsertQuery meetingEditQuery)
-    {
-        int updateMeeting= 0;
-
+    public AjaxResult edit(MeetingInsertQuery meetingEditQuery) {
+        int updateMeeting = 0;
         // 如果传了图片就调用然后更新
-        if (!meetingEditQuery.getFile().isEmpty()) {
-             AjaxResult ajaxResult = remoteCosService.uploadFileCommon(meetingEditQuery.getFile(), meetingEditQuery.getImageId());
-             if (ajaxResult.get("code").toString().equals("200")) {
-                  // 走到这里image库里面插入了一条新的数据，现在删除老的图片
-                  if (meetingEditQuery.getUrl().startsWith("https")) {
-                      // 如果url开头，那么就说明开始的时候插入了数据，需要去Image库里面删除
-                      remoteCosService.removeImage(meetingEditQuery.getUrl());
-                  }
-              }
+        if (meetingEditQuery.getFile() != null ) {
+            if (!meetingEditQuery.getFile().isEmpty()) {
+                AjaxResult ajaxResult = remoteCosService.uploadFileCommon(meetingEditQuery.getFile(), meetingEditQuery.getImageId());
+                if (ajaxResult.get("code").toString().equals("200")) {
+                    // 走到这里image库里面插入了一条新的数据，现在删除老的图片
+                    if (meetingEditQuery.getUrl().startsWith("https")) {
+                        // 如果url开头，那么就说明开始的时候插入了数据，需要去Image库里面删除
+                        remoteCosService.removeImage(meetingEditQuery.getUrl());
+                    }
+                }
+            }
         }
 
         Meeting meeting = new Meeting();
         BeanUtils.copyProperties(meetingEditQuery, meeting);
-        String filename = meetingEditQuery.getFile().getOriginalFilename();
-        String extend = filename.substring(filename.lastIndexOf(".") + 1);
-        if (!meetingEditQuery.getFile().isEmpty())
-            meeting.setUrl("https://jn-1306384632.cos.ap-nanjing.myqcloud.com/common/" + meetingEditQuery.getImageId() + "." + extend);
+        if (meetingEditQuery.getFile() != null) {
+            String filename = meetingEditQuery.getFile().getOriginalFilename();
+            String extend = filename.substring(filename.lastIndexOf(".") + 1);
+            if (!meetingEditQuery.getFile().isEmpty())
+                meeting.setUrl(CosConstant.COS_PATH + "common/" + meetingEditQuery.getImageId() + "." + extend);
+        }
         updateMeeting = meetingService.updateMeeting(meeting);
 
         return toAjax(updateMeeting == 1);
@@ -126,9 +137,8 @@ public class MeetingController extends BaseController
      */
     @RequiresPermissions("meeting:meeting:remove")
     @Log(title = "会议", businessType = BusinessType.DELETE)
-	@DeleteMapping("/{ids}")
-    public AjaxResult remove(@PathVariable Long[] ids)
-    {
+    @DeleteMapping("/{ids}")
+    public AjaxResult remove(@PathVariable Long[] ids) {
         return toAjax(meetingService.deleteMeetingByIds(ids));
     }
 }
